[OPENMP]Provide correct data sharing attributes for loop control variables.
Loop control variables are private in loop-based constructs and we shall
take this into account when generate the code for inner constructs.
Currently, those variables are reported as shared in many cases. Moved
the analysis of the data-sharing attributes of the loop control variable
to an early semantic stage to correctly handle their attributes.

git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk@366474 91177308-0d34-0410-b5e6-96231b3b80d8